Refactor to use asyncio
- add locking for user object to prevent duplicate namespace creation
