+++ test/roles/test_claim_existing/tasks/main.yaml
@@ -1,0 +1,7 @@
+---
+- name: Run tests
+  include_tasks:
+    file: test.yaml
+- name: Cleanup after running tests
+  include_tasks:
+    file: cleanup.yaml
